+++ api/common/defaults.ts
@@ -33,6 +33,8 @@ export const FullConfigDefaults: Partial<Static<typeof FullConfig>> = {
     'scim::enabled': false,
     'proxy::whitelist': [],
     'media::proxy::allow': [],
+    'osm::enabled': false,
+    'osm::url': 'https://nominatim.openstreetmap.org',
     'login::name': 'CloudTAK',
     'login::logo': `data:image/svg+xml;base64,${fs.readFileSync(new URL('../web/public/CloudTAKLogo.svg', import.meta.url)).toString('base64')}`,
     'login::signup': '',